+++ src/lib9/dirfstat.c
@@ -1,3 +1,5 @@
+// +build !plan9
+
 /*
 Plan 9 from User Space src/lib9/dirfstat.c
 http://code.swtch.com/plan9port/src/tip/src/lib9/dirfstat.c
@@ -43,10 +45,10 @@ dirfstat(int fd)
 
 	snprint(tmp, sizeof tmp, "/dev/fd/%d", fd);
 	nstr = _p9dir(&st, &st, tmp, nil, nil, nil);
-	d = malloc(sizeof(Dir)+nstr);
+	d = malloc(sizeof(Dir)+(size_t)nstr);
 	if(d == nil)
 		return nil;
-	memset(d, 0, sizeof(Dir)+nstr);
+	memset(d, 0, sizeof(Dir)+(size_t)nstr);
 	str = (char*)&d[1];
 	_p9dir(&st, &st, tmp, d, &str, str+nstr);
 	return d;
