+++ src/lib9/readn.c
@@ -1,3 +1,5 @@
+// +build !plan9
+
 /*
 Inferno lib9/readn.c
 http://code.google.com/p/inferno-os/source/browse/lib9/readn.c
@@ -36,7 +38,7 @@ readn(int f, void *av, long n)
 	a = av;
 	t = 0;
 	while(t < n){
-		m = read(f, a+t, n-t);
+		m = read(f, a+t, (size_t)(n-t));
 		if(m <= 0){
 			if(t == 0)
 				return m;
